+++ Block/Catalog/Installments.php
@@ -7,28 +7,42 @@
  */
 namespace Afterpay\Afterpay\Block\Catalog;
 
-use Magento\Framework\Registry as Registry;
+use Afterpay\Afterpay\Block\JsConfig;
 use Afterpay\Afterpay\Model\Config\Payovertime as AfterpayConfig;
 use Afterpay\Afterpay\Model\Payovertime as AfterpayPayovertime;
-use Magento\Framework\View\Element\Template\Context;
 use Magento\Framework\Locale\Resolver as Resolver;
+use Magento\Framework\Registry as Registry;
+use Magento\Framework\View\Element\Template\Context;
 
-class Installments extends \Afterpay\Afterpay\Block\JsConfig
+class Installments extends JsConfig
 {
-  
-    protected $registry;
-    protected $afterpayConfig;
-    protected $afterpayPayovertime;
+    /**
+     * @var Registry
+     */
+    private $registry;
+
+    /**
+     * @var AfterpayConfig
+     */
+    private $afterpayConfig;
+
+    /**
+     * @var AfterpayPayovertime
+     */
+    private $afterpayPayovertime;
+
+    /**
+     * @var Resolver
+     */
     private $localeResolver;
 
     /**
-     * Installments constructor.
      * @param Context $context
-     * @param AfterpayConfig $afterpayConfig
-     * @param AfterpayPayovertime $afterpayPayovertime
      * @param Registry $registry
      * @param AfterpayConfig $afterpayConfig
+     * @param AfterpayPayovertime $afterpayPayovertime
      * @param array $data
+     * @param Resolver $localeResolver
      */
     public function __construct(
         Context $context,
@@ -42,63 +56,46 @@ class Installments extends \Afterpay\Afterpay\Block\JsConfig
         $this->afterpayConfig = $afterpayConfig;
         $this->afterpayPayovertime = $afterpayPayovertime;
         $this->localeResolver = $localeResolver;
-        parent::__construct($afterpayConfig,$context, $localeResolver,$data);
+        parent::__construct($afterpayConfig, $context, $localeResolver, $data);
     }
 
     /**
      * @return bool
      */
-    protected function _getPaymentIsActive()
+    public function canShow(): bool
     {
-        return $this->afterpayConfig->isActive();
+        // check if payment is active
+        if ($this->_getPaymentIsActive() &&
+            $this->afterpayConfig->getCurrencyCode() &&
+            $this->afterpayPayovertime->canUseForCurrency($this->afterpayConfig->getCurrencyCode())
+        ) {
+            $excluded_categories = $this->afterpayConfig->getExcludedCategories();
+            if ($excluded_categories != "") {
+                $excluded_categories_array = explode(",", $excluded_categories);
+                $product = $this->registry->registry('product');
+                $categoryids = $product->getCategoryIds();
+                foreach ($categoryids as $k) {
+                    if (in_array($k, $excluded_categories_array)) {
+                        return false;
+                    }
+                }
+            }
+
+            return true;
+        }
+
+        return false;
     }
 
     /**
-     * @return bool
-     */
-    public function canShow()
-    {	
-        // check if payment is active
-        if (!$this->_getPaymentIsActive()) {
-		    return false;
-        }
-		else{
-			if($this->afterpayConfig->getCurrencyCode()){
-				if($this->afterpayPayovertime->canUseForCurrency($this->afterpayConfig->getCurrencyCode())){
-					$excluded_categories=$this->afterpayConfig->getExcludedCategories();
-					if($excluded_categories!=""){
-						$excluded_categories_array =  explode(",",$excluded_categories);
-						$product = $this->registry->registry('product');
-						$categoryids = $product->getCategoryIds();
-						foreach($categoryids as $k)
-						{
-							if(in_array($k,$excluded_categories_array)){
-								return false;
-							}
-						}
-					}
-					return true;				
-				}
-				else{
-					return false;
-				}
-			} 
-			else {
-				return false;
-			}
-		}
-    }	   
-      
-
-	/**
      * @return string
      */
     public function getTypeOfProduct()
     {
         $product = $this->registry->registry('product');
-		return $product->getTypeId();
+        return $product->getTypeId();
     }
-    
+
     /**
      * @return string
      */
@@ -106,13 +103,13 @@ class Installments extends \Afterpay\Afterpay\Block\JsConfig
     {
         // get product
         $product = $this->registry->registry('product');
-        
+
         // set if final price is exist
-        $price = $product->getFinalPrice();
-       
-        return !empty($price)?number_format($price, 2,".",""):"0.00";
-       
-    }  
+        $price = $product->getPriceInfo()->getPrice('final_price')->getValue();
+
+        return !empty($price) ? number_format($price, 2, ".", "") : "0.00";
+    }
+
     /**
      * @return boolean
      */
@@ -120,13 +117,10 @@ class Installments extends \Afterpay\Afterpay\Block\JsConfig
     {
         $canUse=false;
         //Check for Supported currency
-        if($this->afterpayConfig->getCurrencyCode())
-        {
+        if ($this->afterpayConfig->getCurrencyCode()) {
             $canUse= $this->afterpayPayovertime->canUseForCurrency($this->afterpayConfig->getCurrencyCode());
-        } 
-        
+        }
+
         return $canUse;
-        
     }
-    
 }