+++ Model/Adapter/V2/AfterpayOrderTokenV2.php
@@ -44,6 +44,7 @@ class AfterpayOrderTokenV2
     protected $_countryFactory;
     protected $_scopeConfig;
     protected $listStateRequired;
+    private $expressCheckout;
 
     /**
      * AfterpayOrderToken constructor.
@@ -78,6 +79,8 @@ class AfterpayOrderTokenV2
         $this->_countryFactory = $countryFactory;
         $this->_scopeConfig = $scopeConfig;
         $this->listStateRequired = $this->_getStateRequired();
+        $this->expressCheckout=\Afterpay\Afterpay\Model\Config\Source\CartMode::EXPRESS_CHECKOUT;
+        
     }
 
     /**
@@ -89,10 +92,20 @@ class AfterpayOrderTokenV2
      */
     public function generate($object,$override = [])
     {
-        $requestData = $this->_buildOrderTokenRequest($object, $override);
-        $targetUrl = $this->_afterpayConfig->getApiUrl('v2/checkouts/', null, $override);
-        $requestData = $this->constructPayload($requestData, $this->listStateRequired);
-        $this->handleValidation($requestData);
+        if(isset($override["mode"]) && ($override["mode"]==$this->expressCheckout)){
+            $requestData = $this->_buildExpressOrderTokenRequest($object, $override);
+        }else{
+            $requestData = $this->_buildOrderTokenRequest($object, $override);
+        }
+        $targetUrl = $this->_afterpayConfig->getApiUrl('v2/checkouts/', null, $override);       
+        $this->_helper->debug("Request :",$requestData);
+        if(isset($override["mode"]) && ($override["mode"]==$this->expressCheckout)){
+            $this->handleExpressValidation($requestData);
+        }else{
+            $requestData = $this->constructPayload($requestData, $this->listStateRequired);
+            $this->handleValidation($requestData);
+        }
+       
         $response = $this->performTransaction($requestData, $targetUrl);
         return $response;
     }
@@ -183,6 +196,26 @@ class AfterpayOrderTokenV2
         }
     }
 
+    public function handleExpressValidation($requestData)
+    {
+        $errors = [];
+        
+        if (empty($requestData['amount'])) {
+            $errors[] = 'Amount is required';
+        }
+        if (empty($requestData['mode'])) {
+            $errors[] = 'Mode is required';
+        }
+        if(empty($requestData['merchant']['popupOriginUrl'])){
+            $errors[] = "Merchant's origin url is required";
+        }
+        
+        if (count($errors)) {
+            throw new \Magento\Framework\Exception\LocalizedException(__(implode(' ; ', $errors)));
+        } else {
+            return true;
+        }
+    }
 
     /**
      * Build object for order token
@@ -294,6 +327,113 @@ class AfterpayOrderTokenV2
             'currency' => (string)$data['store_currency_code'],
         ];
 
+        return $params;
+    }
+    
+    /**
+     * Build object for Express order token
+     *
+     * @param \Magento\Sales\Model\Order $object Order to get token for
+     * @param $code
+     * @param array $override
+     * @return array
+     */
+    protected function _buildExpressOrderTokenRequest($object, $override = [])
+    {
+        $precision = self::DECIMAL_PRECISION;
+        $data = $object->getData();
+        $billingAddress  = $object->getBillingAddress();
+        $shippingAddress = $object->getShippingAddress();
+        
+        $email = $object->getCustomerEmail();
+        
+        $params['mode'] ="express";
+        
+        $params['merchantReference'] = array_key_exists('merchantOrderId', $override) ? $override['merchantOrderId'] : $object->getIncrementId();
+        
+        $params['merchant'] = [
+            'popupOriginUrl'    => $this->_storeManagerInterface->getStore($object->getStore()->getId())->getBaseUrl() . 'checkout/cart'
+        ];
+        
+        foreach ($object->getAllVisibleItems() as $item) {
+            if (!$item->getParentItem()) {
+                
+                $product = $this->_productRepositoryInterface->getById($item->getProductId());
+                $category_ids = $product->getCategoryIds();
+                $imageHelper =  $this->_objectManagerInterface->get('\Magento\Catalog\Helper\Image');
+                
+                $categories=[];
+                if(count($category_ids) > 0){
+                    foreach($category_ids as $category){
+                        $cat = $this->_objectManagerInterface->create('Magento\Catalog\Model\Category')->load($category);
+                        array_push($categories,$cat->getName());
+                    }
+                }
+                $params['items'][] = [
+                    'name'     => (string)$item->getName(),
+                    'sku'      => (string)$item->getSku(),
+                    'pageUrl'  =>  $product->getProductUrl(),
+                    'imageUrl' =>  $imageHelper->init($product, 'product_page_image_small')->setImageFile($product->getFile())->getUrl(),
+                    'quantity' => (int)$item->getQty(),
+                    'price'    => [
+                        'amount'   => round((float)$item->getBasePriceInclTax(), $precision),
+                        'currency' => (string)$data['store_currency_code']
+                    ],
+                    'categories' => [$categories]
+                ];
+            }
+        }
+        if ($object->getShippingInclTax()) {
+            $params['shippingAmount'] = [
+                'amount'   => round((float)$object->getBaseShippingInclTax(), $precision), // with tax
+                'currency' => (string)$data['store_currency_code']
+            ];
+        }
+        if (isset($data['discount_amount'])) {
+            $params['discounts']['displayName'] = 'Discount';
+            $params['orderDetail']['amount']     = [
+                'amount'   => round((float)$data['base_discount_amount'], $precision),
+                'currency' => (string)$data['store_currency_code']
+            ];
+        }
+        $taxAmount = array_key_exists('base_tax_amount', $data) ? $data['base_tax_amount'] : $shippingAddress->getBaseTaxAmount();
+        $params['taxAmount'] = [
+            'amount'   => isset($taxAmount) ? round((float)$taxAmount, $precision) : 0,
+            'currency' => (string)$data['store_currency_code']
+        ];
+        
+        if(!empty($shippingAddress) && !empty($shippingAddress->getStreetLine(1)))
+        {
+            $params['shipping'] = [
+                'name'          => (string)$shippingAddress->getFirstname() . ' ' . $shippingAddress->getLastname(),
+                'line1'         => (string)$shippingAddress->getStreetLine(1),
+                'line2'         => (string)$shippingAddress->getStreetLine(2),
+                'area1'         => (string)$shippingAddress->getCity(),
+                'area2'         => "",
+                'postcode'      => (string)$shippingAddress->getPostcode(),
+                'region'         => (string)$shippingAddress->getRegion(),
+                'countryCode'   => (string)$shippingAddress->getCountryId(),
+                'phoneNumber'   => (string)$shippingAddress->getTelephone(),
+            ];
+        }
+        if(!empty($billingAddress) && !empty($billingAddress->getStreetLine(1))){
+            $params['billing'] = [
+                'name'          => (string)$billingAddress->getFirstname() . ' ' . $billingAddress->getLastname(),
+                'line1'         => (string)$billingAddress->getStreetLine(1),
+                'line2'         => (string)$billingAddress->getStreetLine(2),
+                'area1'         => (string)$billingAddress->getCity(),
+                'area2'         => "",
+                'postcode'      => (string)$billingAddress->getPostcode(),
+                'region'         => (string)$billingAddress->getRegion(),
+                'countryCode'   => (string)$billingAddress->getCountryId(),
+                'phoneNumber'   => (string)$billingAddress->getTelephone(),
+            ];
+        }
+        $params['amount'] = [
+            'amount'   => round((float)$object->getBaseGrandTotal(), $precision),
+            'currency' => (string)$data['store_currency_code'],
+        ];
+        
         return $params;
     }
 }